+++ user/psh.h
@@ -1,0 +1,17 @@
+// check which line to determine color, rets line number
+int read_line_count();
+
+// updates line num for color
+void save_line_count(int line_count);
+
+// checks if psh is enabled, rets 1 for true, 0 for false
+int is_psh_enabled();
+
+// resets color to white
+void reset_color();
+
+// updates color
+void increment_color();
+
+// sets the color
+void set_color();
